refactor: use python poetry for project management
High-level overview of changes:

1. rename src to flaresolverr in order to namespace the package
   appropriately

2. Move package.json from root folder into flaresolverr so that it can
   be read by the utils.get_flaresolverr_version() method

3. Rewrite necessary import statements to use the full package name as
   per step 1

4. Move script entry point into separate method (cli_run)

5. Manually added poetry dependencies as per existing requirements.txt
   file
